feat: Replace test data with actual yelp data by chaining promises
Used a mixture of ES7 await/async and the Promises provided by the
'request-promise' module to group the promises together. I think
the biggest gotcha was that I had to wrap the .map call for each
request in Promise.all to convert the array of Promises into one
Promise that the calling function could call .then on.