feat: Show 'No data available' when yelpData is missing
Use knockout's conditional 'if' operator to show different content
if yelp data is missing. This happens not only if we don't have anything
for the location at all (i.e. missing 'alias') but also if yelp itself
returns empty fields for any of the requested 'rating', 'review_count',
and 'url'

Contributes to issue #3