feat: Make side bar result highlighted when clicking on marker
This works but still has problems. I implemented it by passing around
a flag that determined whether or not the original action was from the
marker or from the search bar, and if the former than activate the highlight.

Two problems I can see right now: a) if you click on a marker whose matching
side bar result is out of the scroll view, you don't see it expand. To fix this
I'd have to modify the scrolling position of the #results div, and b) there
are duplicate re-renders going around. If I click on the marker after it opens
the side bar result, it keeps the info window open but collapses the side bar result

Contributes to issue #1